# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eventus API",
    "version": "1.0.0",
    "description": "API for creating, reading, updating, and deleting service events."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Service health and statistics"
    },
    {
      "name": "Events",
      "description": "Event management"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Check service health",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "The service is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/event/util/count": {
      "get": {
        "tags": ["Health"],
        "summary": "Get the number of stored events",
        "operationId": "getEventCount",
        "responses": {
          "200": {
            "description": "The current number of events.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "List events",
        "description": "Returns up to five events. Pass `nextId` as a query parameter to continue from a pagination cursor.",
        "operationId": "listEvents",
        "parameters": [
          {
            "name": "nextId",
            "in": "query",
            "required": false,
            "description": "UUID of the event to use as the pagination cursor.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of events.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "description": "The pagination cursor does not identify an event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Next Event not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create an event",
        "operationId": "createEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The event was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/events/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/EventId"
        }
      ],
      "get": {
        "tags": ["Events"],
        "summary": "Get an event",
        "operationId": "getEvent",
        "responses": {
          "200": {
            "description": "The requested event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Events"],
        "summary": "Replace an event",
        "operationId": "updateEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The event was updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        }
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Delete an event",
        "operationId": "deleteEvent",
        "responses": {
          "200": {
            "description": "The event was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteEventResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "EventId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The UUID of the event.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "The request body failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "EventNotFound": {
        "description": "The requested event does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Event not found"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "The server could not validate its generated response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "EventInput": {
        "type": "object",
        "required": ["name", "service", "message", "eventType", "timestamp"],
        "properties": {
          "name": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "eventType": {
            "$ref": "#/components/schemas/EventType"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Event": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EventInput"
          },
          {
            "type": "object",
            "required": ["id"],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ]
      },
      "EventType": {
        "type": "string",
        "enum": ["log", "error"]
      },
      "EventsRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "nextId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "EventsResponse": {
        "type": "object",
        "required": ["events"],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "nextId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok"]
          }
        }
      },
      "CountResponse": {
        "type": "object",
        "required": ["count"],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "DeleteEventResponse": {
        "type": "object",
        "required": ["message", "event"],
        "properties": {
          "message": {
            "type": "string",
            "enum": ["deleted"]
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          }
        }
      },
      "ValidationIssue": {
        "type": "object",
        "additionalProperties": true,
        "required": ["code", "message", "path"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "array",
            "items": {}
          }
        }
      }
    }
  }
}